backend/requirements/base.txt
celery==5.2.7
Django==4.1.6
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
ipython==8.10.0
redis==3.5.3
